+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [#7] add German translations for docs
+### Changed
+- [#11] Add Failed-Condition to mark Errors on debug-modes  
 
 ## [v1.0.0] - 2025-10-01
 ### Added